+++ src/main/java/kitchenpos/menus/infra/NewJpaMenuGroupRepository.java
@@ -1,6 +1,8 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.infra;
 
 
+import kitchenpos.menus.tobe.domain.MenuGroupRepository;
+import kitchenpos.menus.tobe.domain.NewMenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
